requirements.txt
google-genai>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
aiohttp>=3.9.0
asyncio
typing-extensions>=4.0.0
streamlit>=1.28.0